fix: move EncryptionHandlerOptions to OpenKMS namespace
