+++ src/OpenKMS/Providers/EncryptionSchemeProvider.cs
@@ -1,6 +1,5 @@
 using Microsoft.Extensions.Options;
 using OpenKMS.Abstractions;
-using OpenKMS.Options;
 
 namespace OpenKMS.Providers;
